move some utilities to common module
