+++ src/common/mod.rs
@@ -1,0 +1,8 @@
+// Copyright (C) 2022 Alibaba Cloud. All rights reserved.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+//! Some utilities to support fuse-backend-rs.
+
+pub mod file_buf;
+pub mod file_traits;
